+++ scripts/macros/2014/classFeatures/rogue/phantom/wailsFromTheGrave.js
@@ -1,0 +1,5 @@
+// see sneakAttack.js
+export let wailsFromTheGrave = {
+    name: 'Wails from the Grave',
+    version: '1.5.16'
+};